% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeExtractR.R
\name{extractMuSigma}
\alias{extractMuSigma}
\title{Extract mu and sigma}
\usage{
extractMuSigma(beliefs)
}
\arguments{
\item{beliefs}{Takes the output of the updateBeliefs function.}
}
\value{
A list of the mu and sigma of the respective distributions at each
trial for each randomization.
}
\description{
This function extracts the mu and sigma associated with each randomization.
}
