% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeExtractR.R
\name{rand_blocks}
\alias{rand_blocks}
\title{Randomize blocks}
\usage{
rand_blocks(num_rands, Blocks)
}
\arguments{
\item{num_rands}{Number of desired randomizations}

\item{Blocks}{Dataframes of the desired blocks to be randomized, input as c(1,2)}
}
\value{
A list that contains a dataframe for each randomization in the order specified
}
\description{
This function creates a set of randomized blocks for future use. It assumes
only two blocks with a randomized trial order and block order designated by 
the operator.
}
