% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeExtractR.R
\name{se}
\alias{se}
\title{Standard error of the mean}
\usage{
se(x)
}
\arguments{
\item{x}{A series of number with which to calculate the standard error.}
}
\value{
The standard error of the mean.
}
\description{
This function calculates the standard error of the mean 
given a series of numbers.
}
