% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeExtractR.R
\name{selectTrials}
\alias{selectTrials}
\title{Select specific trials}
\usage{
selectTrials(paramUpdates, trials)
}
\arguments{
\item{paramUpdates}{Takes the output of the extractMuSigma function.}

\item{trials}{A list of the trials of interest input as c(1,2,3,etc.).}
}
\value{
A list of the mu and sigma at each designated trial for each
randomization.
}
\description{
This function allows the user to select specific trials of interest.
}
