% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeExtractR.R
\name{slopeExtract}
\alias{slopeExtract}
\title{Extract slope}
\usage{
slopeExtract(trialParameters, VOTspace, boundary)
}
\arguments{
\item{trialParameters}{Takes the output of the selectTrials function.}

\item{VOTspace}{A list of the parameters for the VOT space. Input as
c(1,2,3), where 1 = minimum, 2 = maximum, and 3 = step count.}

\item{boundary}{The designated percentile at with to extract the slope.
This is typically done at the boundary, designated by 0.5. The function
pulls the slopes closest to the designated percentile, which may not 
always be exact.}
}
\value{
A dataframe of the extracted slopes for each randomization at 
each trial. The dataframe inclues the associated VOT, percentile, trial
number and randomization. Also returns three lists of dataframes that 
contains to the probabilities generated by the simplified Bayes rule at
the trials indicated in the selectTrials function.
}
\description{
This function allows the user extract slopes of the internal category
distributions. It first calculates the distributions of the internal
categories along the designated VOT space. Then uses these distributions
to create an indentification function using a simplied Bayes rule. Then
takes the implicit differential along the ID function and pulls the slope
closest to the provided boundary percentile.
}
