% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeExtractR.R
\name{updateBeliefs}
\alias{updateBeliefs}
\title{Update beliefs}
\usage{
updateBeliefs(randomizations, priorParamsG, priorParamsK)
}
\arguments{
\item{randomizations}{Takes a list of randomized trials.}

\item{priorParamsG}{A series of parameters associated with the G distribution.
Entered as c(1,2,3,4). Where 1 = mean, 2 = variance, 3 = mean psuedocount, and
4 = variance psuedocount.}

\item{priorParamsK}{A series of parameters associated with the K distribution.
Entered as c(1,2,3,4). Where 1 = mean, 2 = variance, 3 = mean psuedocount, and
4 = variance psuedocount.}
}
\value{
A list of the randomizations in the same order as presented with an
additional column corresponding to the beliefs updated at each trial.
}
\description{
This function runs the input randomizations through the belief_update function
from David Kleinschmidt's beliefupdatr package. It assumes two perceptual categories
that vary along a single acoustic parameter. These categories are assumed to form
normal distributions.
}
